The numbers of scheduled meetings and dates for each member are now listed on the main character info screen (commonactions.cpp, commondisplay.cpp, includes.h) Window title is now set to game name and version (game.cpp) MSVC project file updated to compile correctly with newer source files, and compile with STL markup by default so Express (free) versions of MSVC will work (game.vcproj) minor: int casting added to one line to get rid of minor compiler warning (interrogation.cpp)
git-svn-id: http://svn.code.sf.net/p/lcsgame/code/trunk@475 a218ab89-6f2a-4548-b3d1-e91a371d7644

+++ src/daily/interrogation.cpp
@@ -1135,7 +1135,7 @@ void tendhostage(Creature *cr, char &clearformess) {
             if(!techniques[TECHNIQUE_RESTRAIN])
                 attack += 5;
 
-            attack += rapport[a->id] * 3;
+            attack += int(rapport[a->id] * 3);
 
             y += 1;
             move(y, 0);
